// yugabyted-ui/ui/src/api/src/models/AssessmentCategoryInfo.ts
// tslint:disable
/**
 * Yugabyte Cloud
 * YugabyteDB as a Service
 *
 * The version of the OpenAPI document: v1
 * Contact: support@yugabyte.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// eslint-disable-next-line no-duplicate-imports
import type { AssessmentIssueInfo } from './AssessmentIssueInfo';


/**
 * Assessment Issues aggrageted based on Issue Category (migration caveat,datatype,etc)
 * @export
 * @interface AssessmentCategoryInfo
 */
export interface AssessmentCategoryInfo  {
  /**
   * 
   * @type {string}
   * @memberof AssessmentCategoryInfo
   */
  category?: string;
  /**
   * 
   * @type {string}
   * @memberof AssessmentCategoryInfo
   */
  category_description?: string;
  /**
   * 
   * @type {AssessmentIssueInfo[]}
   * @memberof AssessmentCategoryInfo
   */
  issues?: AssessmentIssueInfo[];
}